[website] removed tsx components
