+++ website/docs/style/themes/components/base-tokens.jsx
@@ -2,41 +2,20 @@ import React from 'react';
 import styles from './design-tokens.module.css';
 import Input from '@semcore/input';
 import SearchIcon from '@semcore/icon/search/m';
-import designTokens from './design-tokens.json';
+import baseTokens from './base-tokens.json';
 import DataTable from '@semcore/data-table';
-import Link from '@semcore/link';
-import Tooltip from '@semcore/tooltip';
 import Ellipsis, { useResizeObserver } from '@semcore/ellipsis';
-import { brightness } from '@semcore/utils/lib/color';
 import Copy from '@components/Copy';
+import { ColorPreview } from './design-tokens';
 
-export const ColorPreview: React.FC<{ color: string }> = ({ color }) => {
-  if (!color.startsWith('#') && !color.startsWith('rgba(')) return null;
-
-  const needsBorder = React.useMemo(
-    () => color.startsWith('rgba(') || brightness(color) > 230,
-    [color],
-  );
-
-  return (
-    <div
-      style={{
-        background: color,
-        borderColor: needsBorder ? 'black' : 'transparent',
-      }}
-      className={styles.colorPreview}
-    />
-  );
-};
-
-const DesignTokens: React.FC = () => {
+const BaseTokens = () => {
   const [filter, setFilter] = React.useState('');
   const filteredTokens = React.useMemo(
     () =>
-      designTokens.filter(({ name, rawValue, description }) => {
+      baseTokens.filter(({ name, value, description }) => {
         if (!filter) return true;
         if (name?.toLowerCase().includes(filter.toLowerCase())) return true;
-        if (rawValue?.toLowerCase().includes(filter.toLowerCase())) return true;
+        if (value?.toLowerCase().includes(filter.toLowerCase())) return true;
         if (description?.toLowerCase().includes(filter.toLowerCase())) return true;
         return false;
       }),
@@ -61,7 +40,6 @@ const DesignTokens: React.FC = () => {
           <DataTable.Column name="name" children="Token name" ref={nameHeaderRef} />
           <DataTable.Column name="value" children="Value" ref={valueHeaderRef} />
           <DataTable.Column name="description" children="Description" ref={descriptionHeaderRef} />
-          <DataTable.Column name="components" children="Used in" />
         </DataTable.Head>
         <DataTable.Body virtualScroll={{ rowHeight: 45, tollerance: 10 }} h={800}>
           <DataTable.Cell name="name">
@@ -96,20 +74,20 @@ const DesignTokens: React.FC = () => {
                 children: (
                   <Copy
                     copiedToast="Copied"
-                    toCopy={row.rawValue}
-                    title={`Click to copy "${row.rawValue}"`}
+                    toCopy={row[props.name]}
+                    title={`Click to copy "${row[props.name]}"`}
                     trigger="click"
                     className={styles.tokenValueWrapper}
                   >
                     <div className={styles.tokenValue}>
-                      <ColorPreview color={row.computedValue} />
+                      <ColorPreview color={row[props.name]} />
                       <Ellipsis
                         trim="middle"
                         tooltip={false}
                         containerRect={valueHeaderRect}
                         containerRef={valueHeaderRef}
                       >
-                        {row.rawValue}
+                        {row[props.name]}
                       </Ellipsis>
                     </div>
                   </Copy>
@@ -132,57 +110,10 @@ const DesignTokens: React.FC = () => {
               };
             }}
           </DataTable.Cell>
-          <DataTable.Cell name="components">
-            {(props, row) => {
-              if (!row[props.name].length) {
-                return { children: null };
-              }
-
-              if (row[props.name].length === 1) {
-                return {
-                  children: (
-                    <div>
-                      <Link
-                        target="_blank"
-                        href={`/intergalactic/components/${row[props.name][0]}/`}
-                      >
-                        {row[props.name][0]}
-                      </Link>
-                    </div>
-                  ),
-                };
-              }
-
-              return {
-                children: (
-                  <>
-                    <Tooltip>
-                      <Tooltip.Trigger className={styles.usages}>
-                        {row[props.name].length} components
-                      </Tooltip.Trigger>
-                      <Tooltip.Popper>
-                        {row[props.name].map((componentName, index) => (
-                          <React.Fragment key={componentName}>
-                            <Link
-                              target="_blank"
-                              href={`/intergalactic/components/${componentName}/`}
-                            >
-                              {componentName}
-                            </Link>
-                            {index < row[props.name].length - 1 && ', '}
-                          </React.Fragment>
-                        ))}
-                      </Tooltip.Popper>
-                    </Tooltip>
-                  </>
-                ),
-              };
-            }}
-          </DataTable.Cell>
         </DataTable.Body>
       </DataTable>
     </div>
   );
 };
 
-export default DesignTokens;
+export default BaseTokens;
